launcher/container/requirements.txt
coverage==4.5.1
flake8==3.5.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.10.1
Flask-Migrate==2.2.0
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
psycopg2==2.7.4
pyjwt
